+++ projects/worldskills-angular-lib-tester/src/app/app.component.ts
@@ -16,9 +16,11 @@ import { AlertType } from '../../../worldskills-angular-lib/src/lib/alerts/alert
 import { VoteEntry } from '../../../worldskills-angular-lib/src/lib/polls/models/vote-entry';
 import { DateRange } from '../../../worldskills-angular-lib/src/lib/date/date-range';
 import { Track } from '../../../worldskills-angular-lib/src/lib/polls/models/track';
-import { FileThumbnailView } from '../../../worldskills-angular-lib/src/lib/file/file-thumbnail-preview/file-thumbnail-preview.component';
+import { ResourceThumbnail } from '../../../worldskills-angular-lib/src/lib/file/resource-thumbnail/resource-thumbnail.component';
 import { EntityFetchParams } from 'projects/worldskills-angular-lib/src/lib/entity-tree-select/models/entity-tree-fetch-params';
+import { environment } from '../environments/environment';
 
+// TODO: Cleanup, Each demo should be in its' own componennt
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
@@ -39,8 +41,6 @@ export class AppComponent {
     // breadcrumbs
     showHomeItem = true;
     defaultRoute = '/home';
-
-    languages: Language[];
 
     poll: Poll;
     tracks: Track[];
@@ -88,8 +88,8 @@ export class AppComponent {
     ];
 
 
-    fileThumbnailView1: FileThumbnailView;
-    fileThumbnailView2: FileThumbnailView;
+    fileThumbnailView1: ResourceThumbnail;
+    fileThumbnailView2: ResourceThumbnail;
 
     entitySearchParams: EntityFetchParams;
 
@@ -114,9 +114,6 @@ export class AppComponent {
         this.configureLib();
         this.alerts.setAlert('test', AlertType.info, 'Alert!', 'A random alert', false);
 
-        this.languages = [
-            {code: 'en', name: 'English'}
-        ];
         this.isLoggedIn = false;
         this.dropDownMenuItems = [
             { label: 'Sample', url: '/authorized', hidden: false, requireLogin: false, requiredRoles: []}
@@ -156,18 +153,20 @@ export class AppComponent {
 
     private testFileThumbnailPreview(): void {
         this.fileThumbnailView1 = {
-            fileSize: 1000_000,
-            downloadLink: '',
-            filename: 'file1.pdf',
-            thumbnailImageLink: 'assets/images/thumbnail_upload_21599558244800.png'
+            resource_id: 13229,
+            fileSize: 1352705,
+            downloadLink: `${environment.apiBaseUrl}/resources/download/13229/14648/15576?l=en`,
+            filename: 'WSC2022_WSOS37_Landscape_Gardening (1).pdf',
+            thumbnailLink: `${environment.apiBaseUrl}/resources/thumbnail/13229/14648/15576`
         };
 
         this.fileThumbnailView2 = {
-            fileSize: 1000_000,
-            downloadLink: '',
-            filename: 'file2.png',
+            resource_id: 13230,
+            fileSize: 1218597,
+            downloadLink: `${environment.apiBaseUrl}/resources/download/13230/14649/15577?l=en`,
+            filename: 'WSC2022_WSOS09_IT_Software_Solutions_for_Business.pdf',
             // thumbnailImageLink: 'https://picsum.photos/256/256'
-            thumbnailImageLink: 'assets/images/thumbnail_upload_31599558189164.png',
+            thumbnailLink: `${environment.apiBaseUrl}/resources/thumbnail/13230/14649/15577`,
             description: 'here is the description for file 2'
         };
     }
@@ -186,22 +185,22 @@ export class AppComponent {
         });
 
         this.wsi.authConfigSubject.next({
-            loginUrl: 'http://localhost:50300/oauth/authorize',
-            clientId: '7221138f6772',
-            redirectUri: 'http://localhost:4200/home',
-            userinfoEndpoint: 'http://localhost:8081/users/loggedIn?show_child_roles=false&app_code=500',
+            loginUrl: environment.authUrl,
+            clientId: environment.clientId,
+            redirectUri: environment.redirectUrl,
+            userinfoEndpoint: `${environment.apiBaseUrl}/users/loggedIn?show_child_roles=false&app_code=500`,
             oidc: false
         });
 
         this.wsi.httpConfigSubject.next({
             encoderUriPatterns: [],
-            authUriPatterns: ['api.worldskills.show', 'localhost:8081'],
+            authUriPatterns: environment.apiHttpPattern,
         });
 
         this.wsi.serviceConfigSubject.next({
             appCode: [500],
             // apiEndpoint: 'https://api.worldskills.show',
-            apiEndpoint: 'http://localhost:8081'
+            apiEndpoint: environment.apiBaseUrl
         });
     }