+++ projects/worldskills-angular-lib-tester/src/environments/environment.ts
@@ -2,8 +2,14 @@
 // `ng build --prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
+// data required to setup the dedmo app should be placed here
 export const environment = {
-  production: false
+  production: false,
+  apiHttpPattern: ['api.worldskills.show', 'localhost:8080'],
+  apiBaseUrl: 'http://localhost:8081', // port number may need changing ddependding onn your setup
+  authUrl: 'http://localhost:50300/oauth/authorize',
+  clientId: '7221138f6772',
+  redirectUrl: 'http://localhost:4200/home'
 };
 
 /*