+++ projects/worldskills-angular-lib/src/lib/file/resource-thumbnail-container/resource-thumbnail-container.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ResourceThumbnailContainerComponent } from './resource-thumbnail-container.component';
+
+describe('ResourceThumbnailContainerComponent', () => {
+  let component: ResourceThumbnailContainerComponent;
+  let fixture: ComponentFixture<ResourceThumbnailContainerComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ ResourceThumbnailContainerComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ResourceThumbnailContainerComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});