+++ projects/worldskills-angular-lib/src/lib/footer/footer/footer.component.ts
@@ -4,6 +4,7 @@ import { User } from '../../auth/models/user';
 import { GenericUtil } from '../../common/util/generic.util';
 import { MISSING_LANGUAGE_MESSAGE } from '../footer.const';
 import { WorldskillsAngularLibService } from '../../worldskills-angular-lib.service';
+import { LangUtil } from '../../common/util/lang.util';
 
 @Component({
   selector: 'ws-footer',
@@ -74,5 +75,14 @@ export class FooterComponent implements OnInit {
   languagePrompt(): void {
     alert(MISSING_LANGUAGE_MESSAGE(this.supportEmail || this.wsi.defaultSupportEmailAddress));
   }
+
+  // get list of languages - backward compatible for apps that still want to define their own language
+  getLanguages(): Language[] {
+    if (GenericUtil.isNullOrUndefined(this.languages) || this.languages.length === 0) {
+        return LangUtil.getDefaultLanguages();
+    }
+
+    return this.languages;
+}
 
 }