+++ changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 10.25
+
+* 10.25.0 - minor cleanup
+* 10.25.0 - improved asset copy
+
+## 10.24
+
+* 10.24.0 - Translations nonw exist in the lib, backward compatible
+
+## 10.23
+
+* 10.23.0 - update resource thumbnail and add multiple resource thumbnails container 
+
 ## 10.22
 
 * 10.22.3 - update font urls